It works.

+++ lib.cpp
@@ -10774,7 +10774,7 @@ void DrawPanel::OnMouseMovement(wxMouseEvent &event){
             
             
             
-            for(l=0; l<(points_route_list[i]).size()-1; l++){
+            for(l=0; l<((int)((points_route_list[i]).size()))-1; l++){
                 
                 //if the mouse is hovering over one of the points of route #i, I set the background color of route i in listcontrol_routes to a color different from white, to highlight it, and I highlight also the related sight in listcontrol_sights
